Add operator << to EventType

+++ esep_ss2021_Gruppe2_3/src/DemoClient.cpp
@@ -2,13 +2,14 @@
 #include <iostream>
 #include "dispatcher/cnnMngmnt/QnxConnection.h"
 #include "dispatcher/cnnMngmnt/QnxChannel.h"
+#include "dispatcher/Event.h"
 
 DemoClient::DemoClient(const std::string& name) :
         DispatcherClient(name) {
 }
 
 void DemoClient::handle(dispatcher::Event& event) {
-    std::cout << "Demo Client recieved Event:" << std::endl << "Type:" << static_cast<int>(event.type) << std::endl << "payload: " << event.payload << std::endl << std::endl;
+    std::cout << "Demo Client recieved Event:" << std::endl << "Type:" << event.type << std::endl << "payload: " << event.payload << std::endl << std::endl;
 }
 
 DemoClient::~DemoClient() {
